clients/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "commitdb"
version = "0.1.0"
description = "Python Client for CommitDB SQL Server"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "CommitDB Contributors"}
]
keywords = ["database", "sql", "git", "commitdb"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/nickyhof/CommitDB"
Documentation = "https://github.com/nickyhof/CommitDB#readme"
Repository = "https://github.com/nickyhof/CommitDB"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
commitdb = ["lib/*.so", "lib/*.dylib", "lib/*.dll"]

[project.optional-dependencies]
dev = ["pytest>=7.0"]
ibis = ["ibis-framework>=11.0", "pandas>=2.0"]

[project.entry-points."ibis.backends"]
commitdb = "commitdb.ibis_backend"
